Artifacts Building (#2)
The QMK CLI has a nice `json2c` command that will auto generate the keymap.c file for you from the configurator. Copying and pasting this value for now seems to work and including it into the keymap.c file that has the additional features. next step can probably be append out the custom keymap.c file to the generated one.